Refactoring Event APIs Operations

See merge request tas11/showapp!30

+++ src/event/event.service.ts
@@ -2,10 +2,10 @@ import { HttpException, HttpStatus, Injectable, Logger, NotFoundException, Unaut
 import { JwtService } from '@nestjs/jwt';
 import { InjectRepository } from '@nestjs/typeorm';
 import { from, Observable } from 'rxjs';
-import { OrganisationService } from 'src/organisation/organisation.service';
-import { JwtPayload } from 'src/shared/interfaces';
-import { User } from 'src/user/entities/user.entity';
-import { UserService } from 'src/user/user.service';
+import { OrganisationService } from '../organisation/organisation.service';
+import { JwtPayload } from '../shared/interfaces';
+import { User } from '../user/entities/user.entity';
+import { UserService } from '../user/user.service';
 import { CreateEventDto } from './dto/create-event.dto';
 import { UpdateEventDto } from './dto/update-event.dto';
 import { Event } from './entities/event.entity';
@@ -34,12 +34,13 @@ export class EventService {
     this.validateEvent(createEventDto);
     const organisation= await this.organisationService.findById(organisatinId);
     const user=await this.userService.findOne(userRequest);
-
     if(!user) throw new NotFoundException('user not found');
-
     if(!organisation) throw new NotFoundException('organisation not found');
+    if((organisation.user.role==='USER') ||(userRequest!=organisation.user.userId)) throw new UnauthorizedException('Unmatch role for create only organisation');
+
       this.logger.log(`event is created `);
-     return await this.eventRepository.save({...createEventDto,organisation});
+      const eventCode=this.randomString();
+     return await this.eventRepository.save({...createEventDto,eventCode,organisation});
     
   }
 
@@ -49,7 +50,10 @@ export class EventService {
    */
   public async findAll(): Promise<Event[]> {
     this.logger.log(`retrieve all events`);
-    return await this.eventRepository.find({});
+    const events=await this.eventRepository.find({where:[{isDeleted:false}]});
+    if(events.length<0) throw new NotFoundException('not data found ');
+    return events; 
+    
   }
   
   /** Service: search  an event by title
@@ -57,9 +61,9 @@ export class EventService {
    * @param title string
    * @returns Promise<[createEventDto:CreateEventDto ]>
    */
-  public async getByTitle(title:string): Promise<Event[]>{
-    this.logger.log(`searching event by title : ${title}`);
-    return await this.eventRepository.findByTitle(title);
+  public async searchEvent(key:string): Promise<Event[]>{
+    this.logger.log(`searching event by key : ${key}`);
+    return await this.eventRepository.searchEvent(key);
   }
 
   /** Service: retrieve  a particular event
@@ -68,8 +72,12 @@ export class EventService {
    * @returns Observable<{event:Event }>
    */
   findOne(id: string):Observable<Event> {
-    this.logger.log(`get single event by Id ${id}`)
-    return from(this.eventRepository.findOne(id));
+    this.logger.log(`get single event by Id ${id}`);
+    const event=this.eventRepository.finById(id);
+    if(!event) throw new NotFoundException('event not found');
+    return from(this.eventRepository.findOne({
+      where:[{eventId:id}]
+    }));
   }
 
    /** Service: Update particular event 
@@ -77,9 +85,12 @@ export class EventService {
    * @param updateEventDto UpdateEventDto
    * @returns Promise<{event:Event}>
    */
-  public async update(id: string, updateEventDto: UpdateEventDto):Promise<Event> {
-        const event= await this.eventRepository.findOne(id)
+  public async update(id: string, updateEventDto: UpdateEventDto,userRequest:string):Promise<Event> {
+        const event=await this.eventRepository.findOne(id);
+        const org=await this.organisationService.findById(event.organisation.organisationId);
+        
         if(!event) throw new NotFoundException ('event not found ');
+        if((org.user.role==='USER') ||(userRequest!=event.organisation.user.userId)) throw new UnauthorizedException('Unmatch role for update');
         this.validateEvent(updateEventDto);
         this.logger.log(`update event with id :${id} and updated Event ${updateEventDto} :`)
         return this.eventRepository.editProduct(id,updateEventDto);
@@ -89,9 +100,18 @@ export class EventService {
    * @param id string 
    * @returns Observables<{event:Event}>
    */
-  getEventByOrganisationId(id:string){
-    this.logger.log(`retrieve event owned by organisation id :${id}`)
-    return this.eventRepository.find({where:{organisation:id}});
+  async getEventByOrganisationId(id:string){
+    this.logger.log(`retrieve event owned by organisation id :${id}`);
+    const organisation= await this.organisationService.findById(id);
+    if(!organisation) throw new NotFoundException('Organisation not found');
+    if(organisation.user.role!=='ORGANISER') throw new UnauthorizedException('user is not organiser');
+    const event=this.eventRepository.find({where:{organisation:id}});
+    const events=(await event).map((org)=>{
+      if(org.organisation.organisationId===organisation.organisationId){
+        return org;
+      }
+    })
+    return events;
    
   }
 
@@ -99,11 +119,23 @@ export class EventService {
    * @param id string 
    * @returns Promise<{message: string}>
    */
-  public async remove(id: string):Promise<void> {
-    const event= await this.eventRepository.findOne(id)
+  public async remove(id: string,userId:string):Promise<string> {
+     const event= await this.eventRepository.findOne(id)
         if(!event) throw new NotFoundException (`event ${id} not found`);
-        this.logger.log(`remove event by id:${id}`);
-    return  await this.eventRepository.deleteEvent(id);
+        if(event.organisation.user.role==='USER') throw new UnauthorizedException('user is not organiser');
+        const organisations=this.organisationService.getOrganisationsByUserId(userId);
+        const allowed=(await organisations).map((org)=>{
+          if(org.organisationId===event.organisation.organisationId)return org;
+        })
+
+        if(allowed){
+          this.logger.log(`remove event by id:${id}`);
+          await this.eventRepository.deleteEvent(id)
+        return  "deleted success";
+
+        }else{
+          throw new UnauthorizedException('limited grants for actions');
+        }
   }
 
   
@@ -114,11 +146,21 @@ export class EventService {
    *  @param updateEventDto UUpdateEventDto
    * @returns Promise<{event:Event }>
    */
-  public async cancelEvent(id: string, updateEventDto: UpdateEventDto):Promise<Event> {
+  public async cancelEvent(id: string, updateEventDto: UpdateEventDto,userId:string):Promise<Event> {
     const event= await this.eventRepository.findOne(id)
     if(!event) throw new NotFoundException ('event not found');
-    this.logger.log(`cancel event by id:${id}`);
-    return this.eventRepository.cancelEvent(id,updateEventDto);
+    if(event.organisation.user.role==='USER') throw new UnauthorizedException("event must be canceled by it's organiser ")
+    const organisations=this.organisationService.getOrganisationsByUserId(userId);
+        const allowed=(await organisations).map((org)=>{
+          if(org.organisationId===event.organisation.organisationId)return org;
+        })
+        if(allowed){
+          this.logger.log(`cancel event by id:${id}`);
+          return this.eventRepository.cancelEvent(id,updateEventDto);
+
+        }else{
+          throw new UnauthorizedException('limited grants for actions ');
+        }
   }
 
    /** Service:  validate a particular user
@@ -145,6 +187,7 @@ export class EventService {
     const { startHour, endHour, eventStartDate, eventEndDate }=createEventDto;
     const customDate='01/01/2000';
     const currentDate=new Date(Date.now());
+    currentDate.setHours(0,0,0,0);
     const sDate=new Date(eventStartDate);
     const eDate=new Date(eventEndDate);
     const sHour=new Date(customDate.concat(startHour) );
@@ -161,5 +204,24 @@ export class EventService {
     return true;
 
   }
+
+    /** Service:  generate eventCode
+   *
+   * @returns string
+   */
+  private  randomString():string {
+
+		const characters= '0123456789';
+		let str = "";
+		 const createdEmplCode="EV";
+		const mynewCharacters = characters.split('');
+		const generatedCodeLength =4;
+		for (let i = 0; i < generatedCodeLength; i++) {
+		    const index:number = (Math.random() *10);
+		    const newString:string = characters.substring(index, characters.length - 1);
+		    str += mynewCharacters[newString.length];
+		}
+		return createdEmplCode.concat(str);
+	    }
 
 }
